% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{anchors.getConnection}
\alias{anchors.getConnection}
\title{Retrieve an Anchors Connection}
\usage{
\method{anchors}{getConnection}()
}
\value{
Returns an \linkS4class{AnchorsConnection} object.
}
\description{
Attempt to recover an anchors connection.
}
