% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R, R/anchors_dataframe.R, R/package.R
\docType{package}
\name{anchors}
\alias{anchors}
\alias{anchors.data.frame}
\alias{anchors-package}
\title{Create a model explanation function based on training data}
\usage{
anchors(x, model, perturbator = NULL, bins = NULL, p = 0.5,
  coverage_perturbation_count = 1000, target = NULL,
  maxAnchors = NULL, beams = 2L, delta = 0.1, epsilon = 0.1,
  tau = 0.9, tauDiscrepancy = 0.05, initSamples = 10L,
  allowSuboptimalSteps = TRUE, batchSize = 100L, verbose = FALSE,
  ...)

\method{anchors}{data.frame}(x, model, perturbator = NULL, bins = NULL,
  p = 0.5, coverage_perturbation_count = 1000, target = NULL,
  maxAnchors = NULL, beams = 2L, delta = 0.1, epsilon = 0.1,
  tau = 0.9, tauDiscrepancy = 0.05, initSamples = 10L,
  allowSuboptimalSteps = TRUE, batchSize = 100L, verbose = 0, ...)
}
\arguments{
\item{x}{The training data used for training the model that should be
explained.}

\item{model}{The model whose output should be explained}

\item{perturbator}{Pertubator to perturbate the instances to be explained.}

\item{target}{Target column in [x].
Provide this if model is not of class "WrappedModel".}

\item{maxAnchors}{how many features can be maximal anchored?}

\item{beams}{size of the current candidates for beam search}

\item{delta}{The delta value describing the probability of identifying the best arm == confidence}

\item{epsilon}{The maximum tolerated error == tolerance}

\item{tau}{The desired precision an anchor needs to achieve. If no candidate achieves at least this precision, the one with the best precision will be returned}

\item{tauDiscrepancy}{Usually, it is practically infeasible to sample until the mean and the
upper/lower bounds simultaneously fall below or above the tau threshold.
Therefore, this variable may be introduced to control this discrepancy.}

\item{initSamples}{The number of evaluations sampled for each candidate before it gets evaluated by
the best arm identification algorithm. While theoretically, a guarantee that no
candidates get discarded due to too few samples is provided by delta, using this
argument has practical advantages.}

\item{allowSuboptimalSteps}{if set to false, candidates that are returned by the best arm identification get
removed when their precision is lower than their parent's}

\item{verbose}{should anchors communicate additional information while it's running?  Default: FALSE}

\item{...}{Arguments passed on to methods}
}
\value{
Returns an explainer which can be used together with [explain()] to
explain model predictions.
}
\description{
This is the main function of the `anchors` package. It is a factory function
that returns a new function that can be used to explain the predictions made
by black box models. This is a generic with methods for the different data
types supported by anchors.

For this, a model and a perturbation space is required.
To construct the perturbation space, a training set is used.
Also, discretization can be defined to increase explanation quality.

The anchors package provides two categories of important functions:
methods to handle the underlying Anchors Java Server and functions for explanation and rule creation.
}
\section{Handling the underlying anchors java server}{

}

\section{Explaining machine learning models}{

}

