% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{buildBins}
\alias{buildBins}
\title{Helper function to build bins}
\usage{
buildBins(columnIndex, cuts = NULL, currentBins = NULL,
  categoriesPerBinList = NULL, right = F, disc = T)
}
\arguments{
\item{columnIndex}{the index to write the bin to}

\item{cuts}{the cuts to set}

\item{currentBins}{the currently used list of bins}

\item{categoriesPerBinList}{???}

\item{right}{if upper border is to be included}

\item{disc}{true or null, if discretization is to be enabled for this column}
}
\value{
the new list of bins
}
\description{
TODO what is categoriesPerBinList for?
}
