% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{buildDescription}
\alias{buildDescription}
\title{Builds a printable representation of the discretization used for output formatting}
\usage{
buildDescription(bin, cuts, right, short)
}
\arguments{
\item{bin}{the discretization option}

\item{cuts}{the cuts}

\item{right}{right}

\item{short}{short}
}
\value{
the readable description
}
\description{
TODO remove cuts and right as they are included in bin. Also, include non-numeric discretization usungusing $classes
}
