% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{checkBin}
\alias{checkBin}
\title{Checks if the specified undiscretized value is included in the bin}
\usage{
checkBin(value, bin, binIndex)
}
\arguments{
\item{value}{the undiscretized value}

\item{bin}{the discretization options}

\item{binIndex}{the bin index in the list}
}
\value{
True, if value lies within specified discretization option
}
\description{
TODO I don't really know what this function is there for and at first glance it looks buggy to me
}
