% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{create.empty.discretization}
\alias{create.empty.discretization}
\title{Creates an empty discretization which is also used as a default}
\usage{
create.empty.discretization(featureCount)
}
\arguments{
\item{featureCount}{the number of features to create bins for}
}
\value{
an empty discretization which causes the
dataset to not get discretized
}
\description{
Creates an empty discretization which is also used as a default
}
