% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{discretize.data.frame}
\alias{discretize.data.frame}
\title{Title}
\usage{
discretize.data.frame(df, bins)
}
\arguments{
\item{df}{the data.frame on which discretization will be applied upon}

\item{bins}{the discretization bins}
}
\value{
a discretized data.frame
}
\description{
Title
}
