% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{.getConnection}
\alias{.getConnection}
\title{Retrieve an Anchors Connection}
\usage{
.getConnection()
}
\value{
Returns an connection object.
}
\description{
Attempt to recover an anchors connection.
}
