% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R, R/explain_dataframe.R
\name{explain}
\alias{explain}
\alias{explain.data.frame}
\title{Explain model predictions}
\usage{
explain(x, explainer, labels = NULL, ...)

\method{explain}{data.frame}(x, explainer, labels = NULL, ...)
}
\arguments{
\item{x}{New observations to explain, of the same format as used when
creating the explainer}

\item{explainer}{An `explainer` object to use for explaining the observations}

\item{labels}{The specific labels (classes) to explain in case the model is
a classifier. For classifiers either this or `n_labels` must be given.}

\item{...}{Parameters passed on to the `predict_model()` method}
}
\value{
A data.frame encoding the explanations one row per explained
observation. The columns are:

- `model_type`: The type of the model used for prediction.
- `case`: The case being explained (the rowname in `cases`).
- `model_r2`: The quality of the model used for the explanation
- `model_intercept`: The intercept of the model used for the explanation
- `model_prediction`: The prediction of the observation based on the model
  used for the explanation.
- `feature`: The feature used for the explanation
- `feature_value`: The value of the feature used
- `feature_weight`: The weight of the feature in the explanation
- `feature_desc`: A human readable description of the feature importance.
- `data`: Original data being explained
- `prediction`: The original prediction from the model

Furthermore classification explanations will also contain:

- `label`: The label being explained
- `label_prob`: The probability of `label` as predicted by `model`
}
\description{
Once an explainer has been created using the [anchors()] function it can be used
to explain the result of the model on new observations. The `explain()`
function takes new observation along with the explainer and returns a
data.frame with prediction explanations, one observation per row. The
returned explanations can then be visualised in a number of ways, e.g. with
[plot_features()].

This scenario is described by the previously created anchors.tabular.
}
