% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchors.R
\name{lime}
\alias{lime}
\alias{anchors}
\title{Create a model explanation function based on training data}
\usage{
anchors(x, model, perturbator, discX = NULL, target = NULL, ...)
}
\arguments{
\item{x}{The training data used for training the model that should be
explained.}

\item{model}{The model whose output should be explained}

\item{perturbator}{The pertubator used to perturbate the instances to be explained.}

\item{discX}{A discretized training dataset. If none is provided, autodiscretisation
is applied to \code{x}.}

\item{target}{Target column in x. Only provide this if model is not of class "WrappedModel".}

\item{...}{Arguments passed on to methods}
}
\value{
Return an explainer which can be used together with [explain()] to
explain model predictions.
}
\description{
This is the main function of the `anchors` package. It is a factory function
that returns a new function that can be used to explain the predictions made
by black box models. This is a generic with methods for the different data
types supported by lime.
}
