% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePerturbator.R
\name{makePerturbator}
\alias{makePerturbator}
\title{Construct perturbation operator.}
\usage{
makePerturbator(perturbator, supported = getAvailableRepresentations())
}
\arguments{
\item{supported}{[\code{character}]\cr
Vector of names of supported parameter representations. Possible choices:
\dQuote{tabular}, \dQuote{image}.}

\item{function}{[\code{function}]\cr
Actual perturbation function.}
}
\value{
[\code{anchors_perturbator}] Operator object.
}
\description{
Helper function which constructs an perturbation operator.
}
\note{
In general you will not need this function, but rather one of its
deriviatives like \code{\link{makeTabularPerturbator}} or \code{\link{makeImagePerturbator}}.
}
