% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbate.R
\name{perturbate}
\alias{perturbate}
\title{Perturbate an instance}
\usage{
perturbate(perturbationFun, dataset, bins, instance, anchors, p)
}
\arguments{
\item{perturbationFun}{defines how the instances is perturbated. Usually, you
will use \code{\link{makeTabularPerturbator}} to create a perturbation operator}

\item{dataset}{the dataset from which perturbations will be drawn}

\item{bins}{the bins of the discretized dataset}

\item{instance}{the instance to be perturbated}

\item{anchors}{indizes of features that will not be perturbated}

\item{p}{possibility of an a feature to be perturbated. Anchors are not perturbated.}
}
\value{
perturbated instance
}
\description{
Perturbate an instance
}
