% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerturbFun_properties.R
\name{getPerturbFunProperties}
\alias{getPerturbFunProperties}
\title{Get Properties of Pertubation Function}
\usage{
getPerturbFunProperties(perturbFun)
}
\description{
Get Properties of Pertubation Function
}
