% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotExplanations}
\alias{plotExplanations}
\title{Visualizes the explanation result}
\usage{
plotExplanations(explanations, featureNames = NULL, colPal = NULL,
  adjColorTransparent = 1, pdf = NULL)
}
\arguments{
\item{explanations}{the explanation result}

\item{featureNames}{the featureNames}

\item{colPal}{the color pallet}

\item{adjColorTransparent}{the transparency of the colors (alpha)}

\item{pdf}{whether output should be pdf}
}
\value{
the plot object
}
\description{
Visualizes the explanation result
}
