% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printExplanations}
\alias{printExplanations}
\title{#' Prints a human readable IF-THEN sequence for all explanations
#'
#' This function prints a set of IF-THEN rules for each explanation.
#'
#' @param explainer An `explainer` as returned by [anchors()].
#' @param explanation An `explanations` object as returned by [explain()].
#'
#' @return Prints a set of rules
#'
#' @export}
\usage{
printExplanations(explainer, explanations)
}
\description{
#' Prints a human readable IF-THEN sequence for all explanations
#'
#' This function prints a set of IF-THEN rules for each explanation.
#'
#' @param explainer An `explainer` as returned by [anchors()].
#' @param explanation An `explanations` object as returned by [explain()].
#'
#' @return Prints a set of rules
#'
#' @export
}
