% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printInstance}
\alias{printInstance}
\title{Internal helper function}
\usage{
printInstance(i, explainer, explanations)
}
\arguments{
\item{i}{a counter}

\item{explainer}{An `explainer` as returned by [anchors()].}

\item{explanations}{An `explanations` object as returned by [explain()].}
}
\value{
Prints an instance
}
\description{
Internal helper function
}
