% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printRule}
\alias{printRule}
\title{Internal helper function}
\usage{
printRule(i, explainer, explanations)
}
\arguments{
\item{i}{a counter}

\item{explainer}{An `explainer` as returned by [anchors()].}

\item{explanations}{An `explanations` object as returned by [explain()].}
}
\value{
Prints a rule
}
\description{
Internal helper function
}
