% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{provideBin}
\alias{provideBin}
\title{Discretizes an undiscretized value given a discretization option}
\usage{
provideBin(value, bin)
}
\arguments{
\item{value}{the value to be discretized}

\item{bin}{the bin defining how the value is to be discretized}
}
\value{
the discretized value
}
\description{
Discretizes an undiscretized value given a discretization option
}
