% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretization.R
\name{validate.bins}
\alias{validate.bins}
\title{Validates the passed discretization options and standardizes it}
\usage{
validate.bins(bins, length)
}
\arguments{
\item{bins}{the defined bins}

\item{length}{the number of features in the dataset exclusive of target}
}
\value{
the standardized bins
}
\description{
Validates the passed discretization options and standardizes it
}
