% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_dataframe.R, R/package.R
\docType{package}
\name{anchors}
\alias{anchors}
\alias{anchors.data.frame}
\alias{anchors-package}
\title{anchors: This package provides an interface to the java-based implementation of the anchors explanation algorithm for machine learning models.
It enables High-Precision Model-Agnostic Explanations in your favourite statistical computing language.}
\usage{
\method{anchors}{data.frame}(x, model, perturbator = NULL,
  discX = NULL, target = NULL, preprocess = NULL,
  bin_continuous = TRUE, n_bins = 4, quantile_bins = TRUE,
  use_density = TRUE, ...)
}
\arguments{
\item{bin_continuous}{Should continuous variables be binned when making the explanation}

\item{n_bins}{The number of bins for continuous variables if `bin_continuous = TRUE`}

\item{quantile_bins}{Should the bins be based on `n_bins` quantiles or spread evenly over the range of the training data}

\item{use_density}{If `bin_continuous = FALSE` should continuous data be sampled using a kernel density estimation. If not, continuous features are expected to follow a normal distribution.}
}
\description{
The anchors package provides two categories of important functions:
methods to handle the underlying Anchors Java Server and functions for explanation and rule creation.
}
\section{Handling the underlying anchors java server}{

}

\section{Explaining machine learning models}{

}

