% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{printExplanations}
\alias{printExplanations}
\title{Prints a human readable IF-THEN sequence for all explanations}
\usage{
printExplanations(explainer, explanations)
}
\arguments{
\item{explainer}{A `` as returned by [anchors()].}

\item{explanation}{A `data.frame` as returned by [explain()].}
}
\value{
Prints a set of rules
}
\description{
This function prints a set of IF-THEN rules for each explanation.
}
